Выносит функционал с оповещениями в отдельный модуль
